+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 * Bump MSRV to 1.48
+* Remove implementations of `PartialEq`, `Eq`, `PartialOrd`, `Ord`, and `Hash` from the
+  `impl_array_newtype` macro. Users will now need to derive these traits if they are wanted.
 
 # 0.27.0 - 2023-03-15